Move date parsing logic to DateParser
